Create README.md
In IRC chat you have a IGNORE option, this option is not worth as a person has to ignore the other person every time he/she logs in. My suggestion is, can we not convert the same option into ignore by IP so that we don't have to keep ignoring them again and again. Also can we have an option by which we can ignore a person for life time if we have a registered nick. Please let me know what you feel about it and if you can make the desired changes. 

Thanks, waiting for your reply!